PHP: Extract GetRenderProgressResponse into separate file
