+++ packages/lambda-php/src/GetRenderProgressResponse.php
@@ -1,0 +1,18 @@
+<?php
+
+class GetRenderProgressResponse
+{
+    public int $chunks;
+    public bool $done;
+    public float $overallProgress;
+    public string $type;
+    public ?string $outputFile;
+    public int $lambdasInvoked;
+    public int $renderSize;
+    public int $currentTime;
+    public bool $fatalErrorEncountered;
+    public ?int $timeToFinish;
+    public ?string $outBucket;
+    public ?string $outKey;
+    public ?string $bucket;
+}
