+++ packages/lambda-php/src/PHPClient.php
@@ -4,6 +4,7 @@ namespace Remotion\LambdaPhp;
 use Aws\Credentials\CredentialProvider;
 use Aws\Lambda\LambdaClient;
 use Exception;
+use GetRenderProgressResponse;
 use stdClass;
 
 require_once __DIR__ . '/Version.php';
